bot/requirements.txt
python_telegram_bot==13.7
python-dotenv==0.19.0
emoji==1.4.2
mysql-connector-python==8.0.26